Add --list and don't clutter "sacrebleu -h" with the list of all test sets.
The list was included twice in the "sacrebleu -h" output.
The list is getting quite long, so when printed on single line
it is not much readable anyway.
I think a separate --list option is better (it prints 1 test set per line).

However, if anyone (@mjpost) wants to print the list within sacreble -h,
I suggest to use the ArgumentParser's epilog.

I know that `choices=DATASETS.keys()` does also validation
in addition to documentation, but this functionality is already included
(again, with a nicer multi-line listing of available test sets).