Combination of --echo and --origlang works as (could be) expected.
So now, there are two ways how to evaluate e.g. original English subset of wmt18 en-de:

1) Generate the src and ref subsets using
   `sacrebleu -t wmt18 -l en-de --origlang en --echo src > my-orig-en.src`
   `sacrebleu -t wmt18 -l en-de --origlang en --echo ref > my-orig-en.ref`
   Translate only the subset.
   Evaluate it with `sacrebleu my-orig-en.ref < my-orig-en.translation`.
   Note that there is no --origlang in the evaluation.

2) Generate the full test set (i.e. no --origlang in the `sacrebleu --echo src` command).
   Translate the full test set.
   Evaluate it with `sacrebleu -t wmt18 -l en-de --origlang en < my-wmt18.translation`.

So the meaning of --origlang is slightly different in the evaluation and in --echo,
but I think this is intuitive and both scenarios 1 and 2 are useful in different situations:
scenario 1 if you want to focus only on origlang=en during system tuning,
scenario 2 if you have the full test sets already translated and want to analyze the effect of --origlang.