compat: Use 'exp' smoothing for compat.sentence_bleu() similar to CLI
Since the previous commits also updated the default value of 'floor'
smoothing to 0.1 from 0.0, the users using the compat API will already
be experiencing a behavior change in their future sentence BLEU scores.

This commit also prioritizes the behavioral integrity between the API and the
CLI (#98) by changing the default smoothing method of 'sentence_bleu()' to
'exp' similar to CLI. Because of this difference, the scores between the
API and the sacreBLEU utility with -sl option passed, were never
comparable.

The users that want to preserve comparability with their previously
computed sentence BLEU scores, should therefore call this method
with smooth_method='floor' and smooth_value=0.0 in the future.