Multiref bugfix (#204)
WMT21 only has ref:A, and so we created `ref` for backwards compatibility. However, a separate file was created, causing BLEU to use the same reference file twice in eval. This was effectively a noop, but was incorrect. Instead of writing the alias to a second file, we create backwards compatibility by intercepting the 'echo' request.