`--list -t wmt23 -l cs-uk` should print just cs-uk, not other language pairs
When omitting `-l`, `--list` will still print all the language pairs for that test set.

Motivation:
Originally, `--list` showed just the list of language pairs, so there was no reason
to call it with `-l`, but now it lists all the **fields** for a given language pair
and it is relatively slow (it has to parse the XML files), so it makes sense
to restrict the listing to a single language pair only.