modernize packaging and add release automation (#270)
* add a tox environment for testing package builds

* replace setup.py with static pyproject.toml

Use setuptools to package in the same way, but use
pyproject.toml for compatibility with tools
relying on PEP 517.

* dynamically generate version from tags

Use setuptools-scm to get a version from the tag
when released instead of looking for the contents
of a file. Generate a version.py file to replace
the static value in __init__.py.

* add github action to publish releases for tags

This action responds when a tag is pushed to the
main repository to build and upload both sdist and
wheels to pypi.org. More details in
https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions#publishing-to-package-registries